修复：宠物进化
1. 群宠时召唤5000028，5000047的BUG
2. 宠物进化时，手动选择指定宠物进行进化
3. 群宠顺序在上下线、进化时保持不变
4. 不同职业群学习的群宠不同

+++ src/Application.Core/Gameplay/Plugins/PluginManager.cs
@@ -473,7 +473,7 @@ namespace Application.Core.Gameplay.Plugins
             return false;
         }
 
-        public async Task MoreNpcConversation(IChannelClient c, sbyte mode, sbyte type, int selection, string? inputText = null)
+        public async Task MoreNpcConversation(IChannelClient c, sbyte mode, sbyte type, long selection, string? inputText = null)
         {
             if (c.NPCConversationManager != null)
             {
