+++ src/Application.Core/Scripting/Infrastructure/ConversationContext.cs
@@ -1,7 +1,7 @@
 namespace Application.Core.scripting.Infrastructure
 {
 
-    public record TalkMoreAction(sbyte Mode, sbyte type, int selection, string? inputText);
+    public record TalkMoreAction(sbyte Mode, sbyte type, long selection, string? inputText);
 
     public class ConversationInterruptException() : Exception;
     /// <summary>
